Clean up and rebase
